Remove "unpublished_changes" status
This status is essentially the same as "draft". Rather than having two similar statuses, we should
sign post to users that a live version of the step-by-step exists.

We still need to be able to check whether a draft was previously published so that we know which
delete actions can be performed, so the `unpublished_changes?` check needs to stay.

Also updates some of the tests to specifically create draft step by steps, so that the correct status
is returned.

The Filter specs were using `unpublished_changes` to check for what happens when filtering by a
status with nothing assigned. This has been changed to use `published` instead as that required the
least amount of test changes.