Add id to summary-list content section to prevent ambiguity in tests
When adding new instances of summary-list to the page, tests fail due to ambiguity.
Would ideally add this ID in the same commit as a new summary-list instance is added,
but have a few streams of work going on in tandem so it is useful to be able to
cherry-pick a single commit rather than repeat the fix in multiple places.