Generate a change note when in review
Internal change notes expect the author to be a string of the users name. It currently
doesn't do a lookup on the uid to get the name of the user, so we need to pass it the "name"
of current_user.

Also DRYs-up the status in the controller action to limit the possibility of spelling errors in
automated change notes.