Add a helper method for if user can claim review
This method will be used by the view to determine whether to show a
"claim review" button to the user.

A new view helper has been added because it is expected that this will
be called from the view, and pass in the values for `@step_by_step_page`
and `current_user`. As `current_user` is a controller method, it is not
appropriate to to put this method in the model.