DRY-up tests for review controller
Rather than having an individual test for various combinations of permissions,
we now pre-define the required permissions for each action then loop through
and test removing each.

Note that `drop(1)` is present since we do not want to test removing the
`signin` permission.  Removing this permission causes the users to be
redirected to Signon, rather than being denied access.