Remove margin-top for inset text
The inset text component is used in app with the role of flagging required actions before publishing. It sits the top of the summary list and it should be aligned with the actions block. In order to achieve that we need to remove the margin-top. We ussually only set margins at the bottom of each component - this one being an exception as it's expected to sit in a 'prose' context, surrounded by text.