Fix bug where new steps were added to the top of the step by step
Creating a new step sets a position of `nil` by default. If you have
already created steps and set their positions by reordering them,
then the newly added step is displayed at the top as they are rendered
by `position ASC`, and `nil` evaluates to `0`, thus appearing first.
We now set a step's `position` after it has been created.

NB, had to write the integration test with `js: true` as currently
it is not possible to reorder steps without JavaScript. There is a
card on the backlog to change that:
https://trello.com/c/TSXgNbVK/186-unable-to-reorder-steps-without-javascript