Use translations file to store humanised statuses
This lays the groundwork to have the presentation of a status be
different from the result of calling `.humanize` on a status string.

I decided to use the translations as that seemed an appropriate place
for view concerns. As there is not specific views for this data I put in
a grouping named after the model.