Rename "In review" to "In 2i review"
This changes the public presentation of the in review status to include
information for 2i review.

Ideally we'd also rename the value in the database for this however this
is a delicate operation that could cause some disruption for users so we
should be sure on the value of that before proceeding.