Generate breadcrumbs and headings consistently
This PR updates a series of views to consistently use the admin_layout file for placing breadcrumbs, title and context. This improves the generated markup and makes the copy consistent across pages: the copy for each section we show on the summary page is played back in the heading on the edit pages.