Add hidden_search_terms to publishing-api payload
Trello: https://trello.com/c/JkaMKLxT
Zendesk: https://govuk.zendesk.com/agent/tickets/4167533

The content on coronavirus landing and hub pages isn't being indexed in
the site search.

Search has a concept of `indexable_content`. This means that it looks
for [certain fields](https://github.com/alphagov/search-api/blob/fb3eafcb2edbda5a1b7aeb90d6465342910bcf00/lib/govuk_index/presenters/indexable_content_presenter.rb#L3)
like `body` in the "details" hash of a content item and makes them
searchable.

Unfortunately, the [coronavirus landing](https://www.gov.uk/api/content/coronavirus)
and hub pages have a lot of attributes in "details" but none of them
match the fields that the search-api is looking for.

One of the accepted fields is "hidden_search_terms".
[smart-answers](https://www.gov.uk/api/content/am-i-getting-minimum-wage)
already take advantage of "hidden_search_fields" to index content that
appears in the question flow.

This change goes through all of the section/accordion text content on
the coronavirus page and "flattens" it into a list of sentences and adds
them to "hidden_search_terms".

For example, if the `sections` in the details hash looks like:

```
"sections" => [
  {
    "title" => "Blue Shadow Virus",
    "sub_sections" => [
      {
        "list" => [
          {
            "url" => "/foo",
            "label" => "Apply for a test",
          },
          {
            "url" => "/bar",
            "label" => "What to do if you test positive",
          },
        ],
        "title" => "Testing for Blue Shadow Virus",
      },
    ],
  },
]
```

Then `hidden_search_terms will be populated with:

```
[
  "Blue Shadow Virus",
  "Testing for Blue Shadow Virus",
  "Apply for a test",
  "What to do if you test positive"
]
```

A safe navigation operator (&) is being used when looping through the
"lists" in subsections as the code currently allows users to add content
with just headings in it, i.e. no links.

At the moment these changes only apply to the accordion content, but
they can be expanded to include the other sections later.

The landing and hub pages will need to be republished for these changes
to take effect.