Add timeline content to hidden_search_terms
Add timeline content to hidden_search_terms in publishing-api payload
This allows the timeline content to be indexed by search-api and make it
findable by users using site search.

Timeline content is in the `details` hash in the content item and takes
the following format:

```
timeline: {
  list: [
    {
      heading: "18 September",
      paragraph: "If you live, work or travel in the North East, you need to [follow different covid rules](/guidance/north-east-of-england-local-restrictions)"
    },
    {
      heading: "15 September",
      paragraph: "If you live, work or visit Bolton, you need to [follow different covid rules](/guidance/bolton-local-restrictions)"
    }
  ],
  heading: "Recent and upcoming changes"
}
```
The paragraph values should be appended to the `hidden_search_terms`
list in the content item:

```
hidden_search_terms: [
  .
  .
  .
  "If you live, work or travel in the North East, you need to follow different covid rules",
  "If you live, work or visit Bolton, you need to follow different covid rules"
]
```

Link markdown and url is being stripped out using the markdown helper
built in a previous commit.

The landing and page will need to be republished for these changes to
take effect.