Wrap reordering of announcements in transaction
Wrap the reorder action in a transaction so that it can be rolled back
if the call to publishing-api fails. This has the benefit of preserving
the record, including the current position of the announcement.

The tests have been tied a little. A `let!` is being used to create the
announcement so that it is perpetuated in the tests that need it.