Change namespace of coronavirus controllers
This continues the work to make Coronavirus code distinct in this
application by applying a Coronavirus namespace to coronavirus related
code.

I've used the approach of defining the controller classes within a
module (rather than class Coronavirus::AnnouncementsController) so that
the code in the classes is scoped to the module allowing us to avoid
adding `Coronavirus::` any time we reference a class.