Fix remaining Coronavirus namespace problems
This adds the Coronavirus namespace to the places in the code that
currently break without it. This is done as a quick pass to get the code
working with intentions to do later refactoring to the presenters and
services that make use of Coronavirus code so that they are within the
namespace and no longer require this change.