Add action URL, content and summary to sub section
Before publishing an article, content designers can create drafts of the proposed content. Sometimes, these drafts do not make it to the published stage and are discarded by the user.

When this happens a new draft - from the original live content - needs to be created. To do this the publishing-api is called to get the current live version of the content. The draft discarder is then called which removes any existing sub sections from the current draft and recreates the sub sections based on the live version content.

The draft discarder calls the sections presenter, which in turn calls the sub section processor. Ultimately, a data structure in the correct format is returned that contains the recreated title and sub sections in the format expected by UI.

We need to update this call chain to include the new action_link_url, action_link_content and action_link_summary data (if present) and present that back to the UI for display in the action link view fields.

There are three places that need updating to achieve this:

- the sub section processor,

- the sections presenter, and

- the draft discarder

This first change adds the action_link_content and action_link_summary fields to each sub section that is return by the sub section processor.

In cases where the live version has a previously created featured_link, we need to construct the appropriate action link data within the sub section instead of the original featured_link markdown as previously.